abstract a 'popper' edit cell
