+++ src/features/PantryItemAdmin/components/LabelsEditCell.tsx
@@ -3,28 +3,15 @@ import { Result } from "../../../data/hooks/usePantryItemSearch";
 import { useGetAllLabels } from "../../../data/hooks/useGetAllLabels";
 import { ChipPicker } from "../../../global/components/ChipPicker";
 import * as React from "react";
-import { useCallback, useLayoutEffect, useState } from "react";
-import { Paper } from "@mui/material";
-import Popper from "@mui/material/Popper";
+import { useCallback } from "react";
+import PopperEditCell from "./PopperEditCell";
 
 export default function LabelsEditCell(
     props: GridRenderEditCellParams<Result, string[]>,
 ) {
-    const { id, field, value, colDef, hasFocus } = props;
+    const { id, field, value } = props;
     const { data: labelList } = useGetAllLabels();
-    const [anchorEl, setAnchorEl] = useState<HTMLElement | null>();
-    const [inputRef, setInputRef] = useState<HTMLInputElement | null>(null);
     const apiRef = useGridApiContext();
-
-    useLayoutEffect(() => {
-        if (hasFocus && inputRef) {
-            inputRef.focus();
-        }
-    }, [hasFocus, inputRef]);
-
-    const handleRef = useCallback((el: HTMLElement | null) => {
-        setAnchorEl(el);
-    }, []);
 
     const handleChange = useCallback(
         (e, labels: string[]) => {
@@ -35,38 +22,18 @@ export default function LabelsEditCell(
     );
 
     return (
-        <div style={{ position: "relative", alignSelf: "flex-start" }}>
-            <div
-                ref={handleRef}
-                style={{
-                    height: 1,
-                    width: colDef.computedWidth,
-                    display: "block",
-                    position: "absolute",
-                    top: 0,
-                }}
-            />
-            {anchorEl && (
-                <Popper open anchorEl={anchorEl} placement="bottom-start">
-                    <Paper
-                        elevation={1}
-                        sx={{
-                            p: 1,
-                            minWidth: colDef.computedWidth,
-                            maxWidth: colDef.computedWidth * 2,
-                        }}
-                    >
-                        <ChipPicker
-                            value={value || []}
-                            options={labelList}
-                            size={"small"}
-                            fieldPlaceholder={"Add"}
-                            onChange={handleChange}
-                            inputRef={(ref) => setInputRef(ref)}
-                        />
-                    </Paper>
-                </Popper>
+        <PopperEditCell
+            {...props}
+            renderControl={(inputRef) => (
+                <ChipPicker
+                    value={value || []}
+                    options={labelList}
+                    size={"small"}
+                    fieldPlaceholder={"Add"}
+                    onChange={handleChange}
+                    inputRef={inputRef}
+                />
             )}
-        </div>
+        />
     );
 }
