+++ src/features/PantryItemAdmin/components/MultilineEditCell.tsx
@@ -1,31 +1,18 @@
 import { GridRenderEditCellParams, useGridApiContext } from "@mui/x-data-grid";
 import { Result } from "../../../data/hooks/usePantryItemSearch";
 import * as React from "react";
-import { ChangeEvent, useCallback, useLayoutEffect, useState } from "react";
-import { Paper } from "@mui/material";
-import Popper from "@mui/material/Popper";
+import { ChangeEvent, useCallback } from "react";
 import Input from "@mui/material/Input";
+import PopperEditCell from "./PopperEditCell";
 
 export default function MultilineEditCell(
     props: GridRenderEditCellParams<Result, string[]>,
 ) {
-    const { id, field, value, colDef, hasFocus } = props;
+    const { id, field, value, colDef } = props;
     const [valueState, setValueState] = React.useState(
         value ? value.join(", ") : "",
     );
-    const [anchorEl, setAnchorEl] = useState<HTMLElement | null>();
-    const [inputRef, setInputRef] = useState<HTMLInputElement | null>(null);
     const apiRef = useGridApiContext();
-
-    useLayoutEffect(() => {
-        if (hasFocus && inputRef) {
-            inputRef.focus();
-        }
-    }, [hasFocus, inputRef]);
-
-    const handleRef = useCallback((el: HTMLElement | null) => {
-        setAnchorEl(el);
-    }, []);
 
     const handleChange = useCallback(
         (e: ChangeEvent<HTMLTextAreaElement>) => {
@@ -49,37 +36,18 @@ export default function MultilineEditCell(
     );
 
     return (
-        <div style={{ position: "relative", alignSelf: "flex-start" }}>
-            <div
-                ref={handleRef}
-                style={{
-                    height: 1,
-                    width: colDef.computedWidth,
-                    display: "block",
-                    position: "absolute",
-                    top: 0,
-                }}
-            />
-            {anchorEl && (
-                <Popper open anchorEl={anchorEl} placement="bottom-start">
-                    <Paper
-                        elevation={1}
-                        sx={{
-                            p: 1,
-                            minWidth: colDef.computedWidth,
-                            maxWidth: colDef.computedWidth * 2,
-                        }}
-                    >
-                        <Input
-                            multiline
-                            value={valueState}
-                            placeholder={"Comma-delimited " + colDef.headerName}
-                            onChange={handleChange}
-                            inputRef={(ref) => setInputRef(ref)}
-                        />
-                    </Paper>
-                </Popper>
+        <PopperEditCell
+            {...props}
+            renderControl={(inputRef) => (
+                <Input
+                    multiline
+                    fullWidth
+                    value={valueState}
+                    placeholder={"Comma-delimited " + colDef.headerName}
+                    onChange={handleChange}
+                    inputRef={inputRef}
+                />
             )}
-        </div>
+        />
     );
 }
